Re-write MousepadSettings to use multiple child schemas
For view settings, window settings, search state, and window state.
MousepadSettings no longer is a GSettings subclass but rather contains
a GSettings instance for each of the new child schemas.

Also add constants MOUSEPAD_SETTING_* and MOUSEPAD_STATE_* to use
in code in order to reduce changes of a typo slipping by unnoticed.