Change path parsing code to give caller 0-terminated string
Before, it never used the key length so if the path had a trailing
slash it could be used as part of the key. Now the caller gets the
key name, zero-terminated, into the buffer they provide, and it will
only contain the parsed key (assuming the caller's buffer was big
enough).