Refactor GSettings path/schema code into separate class
It looks through the main GSettings to find all the keys and recurses
into child schemas to find their keys and it makes up a lookup hash
table that maps key paths to their key name and GSettings instance.

The new code is no longer hardcoded to specific schemas/paths except
for the macros that save typing mistakes for the paths.