Remove mutex locking around global settings object
Not using threads at this point anyway and using it at this level
causes deadlocks when functions watching for settings changes call
back in to get/set the settings. In the future if threads are used,
access to the global settings instance should be locked by the caller
or some other mechanism put in places.