Change keyboard shortcuts for Find Next/Previous and Go To
Hopefully this doesn't annoy anyone but these are the more usual
keyboard shortcuts for these uses. See bug report for details. If
anyone has an idea for how to provide both the old and new keyboard
shortcuts without ugly hacks, we could try and support the old and
new shortcuts together.

Closes #9745 (https://bugzilla.xfce.org/show_bug.cgi?id=9745)