Use normal signal connection instead of GClosures
For the proxy menu item connection that allows the statusbar to show
the menu items tooltips on hover. This gets rid of tons of console spam
like this:

    (mousepad:8256): GLib-GObject-CRITICAL **:
    g_closure_add_invalidate_notifier: assertion 'closure->n_inotifiers
    < CLOSURE_MAX_N_INOTIFIERS' failed

I guess we exceeded the hardcoded limit (CLOSURE_MAX_N_INOTIFIERS) or
something. Using regular g_signal_connect*() seems to get rid of all
these warnings and is less code too.