Support statusbar tooltips for toolbar items too
Refactor the code so that the statusbar can itself display tooltips for
widgets and the window code just uses this. For the proxy connections,
tool items need a bit of special treatment since they don't have the
"select" and "deselect" signals used by menu items, so use the Gdk
event signals for mouse enter/leave and focus enter/leave (to cover
mouse and keyboard events) are used on the tool items' child widgets.

Finishes one of the TODOs in 2b51559.