Move GSV language-related code to a new file
Re-write related code so that there's a single action group for the
filetype languages that is shared between any menus to select the
language. Add code to create the menus and use it for the main menu
and the statusbar menu. This way when the language changes, it's
reflected everywhere in the UI automatically.