Fix a bunch of minor compiler warnings
The flags used were:

    -Wall -Wextra -Wno-unused-parameter -Wno-overlength-strings
    -std=c99 -pedantic

The `-Wno-overlength-strings` is needed because we use a tool to convert
text files into C strings that are compiled into the program. Strictly
speaking this is non-standard behaviour as the standard defines a maximum
length of a string literal (probably from the old days when compilers didn't
have enough memory available to handle them?).

Most of the "fixes" were extremely trivial:

* Don't forward declare enums with typedef and then define them.
* Remove semi-colon after `G_DEFINE_TYPE()` since after the macro is expanded
  it leaves a stray semi-colon in the file.
* Add wrapper macro+function `mousepad_disconnect_by_func` to hide the fact
  that GLib's macros are passing our `GCallback` as a gpointer, which is not
  valid C but happens to work most of the time. The new function uses a union
  hack (having same issues as existing) to trick the compiler and a
  function-like wrapper macro to cast the callback function parameter to
  `GCallback` so we still don't need to put `G_CALLBACK()` casts everywhere
  in the code that uses it.

There's still quite a few warnings left but they are not in our code or are
in our generated code. For example with my version of headers Pango is using
values outside the range of `int` as enum values and the generated code for
`mousepad-marshal.c` and `mousepad-dbus-infos.h` is treating function
pointers like void pointers as mentioned above.