Add pkexec policy (Bug #11088)
This adds a pkexec policy file to mousepad. This way if the user
of a desktop system wants to use mousepad to edit a file as root
and has the proper credentials they can.
pkexec comes from polkit which does about the same thing as
sudo/gtksu except also supports authentication methods like
fingerprint or smart card readers.