Prevent a weird cycle in GSettings use
Was causing it to continuously update the "color-scheme" setting,
resulting in lots of DBus spam and CPU/disk activity. I'm not sure
why only this setting was causing the issue, and why it still updates
across all the instances while using the `G_SETTINGS_BIND_GET_NO_CHANGES`
flag, since it shouldn't (and doesn't if it's used on other settings).

Closes bug #12134